Replace non-alphanumeric characters with dash in id attributes.
Although it is allowed to use just about any character for id attributes there are some that will create problems when using them from css or javascript/jquery. For shins it can be seen by adding f.ex. a "." to one of the ids in source/index.html.md. This PR cleans up the id attributes and replaces all non-alpanumeric character with a dash.

See discussion here: https://stackoverflow.com/a/79022